added load balance weight function to the tcpClient. Once a module register to the RC center, it will tell the RC the current weight it has . and when the client get the module instance list, it will get the weight info as well

+++ xixiRouter/src/main/java/xixi/router/AbstractRouter.java
@@ -4,13 +4,13 @@ import xixi.transport.client.TcpClient;
 
 public abstract class AbstractRouter implements Router{
 
-	private int moduleId;
+	private short moduleId;
 	
-	public AbstractRouter(int moduleId){
+	public AbstractRouter(short moduleId){
 		this.moduleId = moduleId;
 	}
 	
-	public int moduleId(){
+	public short moduleId(){
 		return moduleId;
 	}
